requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
crispy==0.6.3
cycler==0.11.0
Django==4.0.4
django-crispy-forms==1.14.0
django-filter==21.1
djangorestframework==3.13.1
fabio==0.13.0
fonttools==4.33.3
h5py==3.6.0
idna==3.3
install==1.3.5
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
silx==1.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
